+++ packages/CB-serverless-frontend/src/sagas/index.js
@@ -4,11 +4,11 @@ import verifyUserSaga from './verifyUserSaga';
 import forgotPasswordRequestSaga from './forgotPasswordRequestSaga';
 import forgotPasswordSaga from './forgotPasswordSaga';
 import requestVerificationCodeSaga from './requestVerificationCodeSaga';
-import cartItemsFetchSaga from './cartItemsFetchSaga';
-import cartItemsAddSaga from './cartItemsAddSaga';
-import cartItemsDeleteSaga from './cartItemsDeleteSaga';
-import cartItemUpdateQtySaga from './cartItemsUpdateQtySaga';
-import cartItemsCleanSaga from './cartItemsCleanSaga';
+import cartItemsFetchSaga from '../modules/Cart/sagas/cartItemsFetchSaga';
+import cartItemsAddSaga from '../modules/Cart/sagas/cartItemsAddSaga';
+import cartItemsDeleteSaga from '../modules/Cart/sagas/cartItemsDeleteSaga';
+import cartItemUpdateQtySaga from '../modules/Cart/sagas/cartItemsUpdateQtySaga';
+import cartItemsCleanSaga from '../modules/Cart/sagas/cartItemsCleanSaga';
 import placeOrderSaga from './placeOrderSaga';
 import cleanOrderSaga from './cleanOrderSaga';
 import paymentTokenIdSubmitSaga from './paymentTokenIdSubmitSaga';
